refactor: add & apply use-loading-block-modal hook
