+++ src/features/arcade-record-article/record-form.tsx
@@ -13,8 +13,7 @@ import { ArcadeRecordPost } from '^/src/entities/types/post';
 import MultipleImagePicker from '^/src/shared/image-picker/multiple';
 import SingleImagePicker from '^/src/shared/image-picker/single';
 import { ImageListElementValue } from '^/src/shared/image-picker/types';
-import { useModalStore } from '^/src/shared/modal/store';
-import { ModalType } from '^/src/shared/modal/types';
+import { useLoadingBlockModal } from '^/src/shared/modal/loading-block';
 import {
   FailedRouteHandlerCallResponse,
   RouteHandlerCallResponse,
@@ -38,11 +37,12 @@ export default function RecordForm({
   methodList,
 }: Props) {
   const route = useRouter();
-  const setModal = useModalStore((state) => state.setModal);
 
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [isSuccess, setIsSuccess] = useState<boolean>(false);
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
+
+  useLoadingBlockModal(isLoading);
 
   const arcadeRecordId = useRef<string>(
     post?.arcadeRecordId ?? uuidv4()
@@ -157,24 +157,6 @@ export default function RecordForm({
       });
     }
   }, [errorMessage]);
-
-  useEffect(() => {
-    if (isLoading) {
-      setModal({
-        type: ModalType.LOADING_BLOCK,
-      });
-    } else {
-      setModal({
-        type: ModalType.OFF,
-      });
-    }
-
-    return () => {
-      setModal({
-        type: ModalType.OFF,
-      });
-    };
-  }, [isLoading, setModal]);
 
   async function handleOnSubmit(event: FormEvent<HTMLFormElement>) {
     event.preventDefault();
