+++ src/features/auth/sign-in-form.tsx
@@ -1,14 +1,13 @@
 'use client';
 
-import { useActionState, useEffect, useState } from 'react';
+import { useActionState, useState } from 'react';
 
 import FormInput from '^/src/shared/ui/form-input';
 
 import { checkIsEmailValid } from '^/src/shared/lib/email';
 import { checkIsPasswordValid } from '^/src/shared/lib/password';
-import { useModalStore } from '^/src/shared/modal/store';
-import { ModalType } from '^/src/shared/modal/types';
 
+import { useLoadingBlockModal } from '^/src/shared/modal/loading-block';
 import { AuthActionState } from './action-state';
 import { signInAction } from './sign-in-action';
 
@@ -17,7 +16,7 @@ export default function SignInForm() {
     AuthActionState,
     FormData
   >(signInAction, {});
-  const setModal = useModalStore((state) => state.setModal);
+  useLoadingBlockModal(isLoading);
 
   const [email, setEmail] = useState<string>('');
   const [password, setPassword] = useState<string>('');
@@ -26,24 +25,6 @@ export default function SignInForm() {
     !isLoading &&
     checkIsEmailValid(email).isPass &&
     checkIsPasswordValid(password).isPass;
-
-  useEffect(() => {
-    if (isLoading) {
-      setModal({
-        type: ModalType.LOADING_BLOCK,
-      });
-    } else {
-      setModal({
-        type: ModalType.OFF,
-      });
-    }
-
-    return () => {
-      setModal({
-        type: ModalType.OFF,
-      });
-    };
-  }, [isLoading, setModal]);
 
   return (
     <form
