+++ src/features/auth/sign-out-form.tsx
@@ -1,33 +1,13 @@
 'use client';
 
-import { useActionState, useEffect } from 'react';
+import { useActionState } from 'react';
 
-import { useModalStore } from '^/src/shared/modal/store';
-import { ModalType } from '^/src/shared/modal/types';
-
+import { useLoadingBlockModal } from '^/src/shared/modal/loading-block';
 import { signOutAction } from './sign-out-action';
 
 export default function SignOutForm() {
   const [, formAction, isLoading] = useActionState(signOutAction, null);
-  const setModal = useModalStore((state) => state.setModal);
-
-  useEffect(() => {
-    if (isLoading) {
-      setModal({
-        type: ModalType.LOADING_BLOCK,
-      });
-    } else {
-      setModal({
-        type: ModalType.OFF,
-      });
-    }
-
-    return () => {
-      setModal({
-        type: ModalType.OFF,
-      });
-    };
-  }, [isLoading, setModal]);
+  useLoadingBlockModal(isLoading);
 
   return (
     <form action={formAction}>
