+++ src/features/review-article/delete-form.tsx
@@ -1,12 +1,10 @@
 'use client';
 
-import { useActionState, useEffect } from 'react';
+import { useActionState } from 'react';
 
 import { ReviewPost } from '^/src/entities/types/post';
 
-import { useModalStore } from '^/src/shared/modal/store';
-import { ModalType } from '^/src/shared/modal/types';
-
+import { useLoadingBlockModal } from '^/src/shared/modal/loading-block';
 import { deleteReviewAction } from './delete-review-action';
 
 interface Props {
@@ -15,25 +13,7 @@ interface Props {
 
 export default function DeleteReviewForm({ reviewId }: Props) {
   const [, formAction, isLoading] = useActionState(deleteReviewAction, null);
-  const setModal = useModalStore((state) => state.setModal);
-
-  useEffect(() => {
-    if (isLoading) {
-      setModal({
-        type: ModalType.LOADING_BLOCK,
-      });
-    } else {
-      setModal({
-        type: ModalType.OFF,
-      });
-    }
-
-    return () => {
-      setModal({
-        type: ModalType.OFF,
-      });
-    };
-  }, [isLoading, setModal]);
+  useLoadingBlockModal(isLoading);
 
   return (
     <form
